fix-me/Nix: reset process-compose field. Todo - confirm its value.
After the upgrade, the process-compose-flake module API seems to have
changed and it doesn't recognise the old fields/syntax.

This commit resets the value of "process-compose" field
to an empty attr set, pending further confirmation on what
it's value should be, since we already have an explicit
"packages.default" value.